+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added headers to publisher
+- Added `assertEmitted` to `Pigeon::fake()`
+- Added `assertConsumingEvent` to `Pigeon::fake()`
+- Added `dispatchListener` to `Pigeon::fake()`
 ### Change
 - Default timeout from 5 to 0
 ### Removed